requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
deep-translator>=1.11.0
openai>=1.0.0
scikit-learn>=1.1.0
ipython>=8.0.0
